+++ src/store/index.ts
@@ -2,7 +2,7 @@ import Vue from "vue";
 import Vuex from "vuex";
 import { version, authors, dependencies } from "../../package.json";
 import players from "@/assets/json/players.json";
-import magicAbilities from "@/assets/json/magic-abilities.json";
+import mtgAbilities from "@/assets/json/mtg-abilities.json";
 import scytheFactions from "@/assets/json/scythe-factions.json";
 import scythePlays from "@/assets/json/scythe-plays.json";
 import tapestryCivilizations from "@/assets/json/tapestry-civilizations.json";
@@ -34,7 +34,7 @@ export default new Vuex.Store({
 
     players: players as Player[],
 
-    magicAbilities: magicAbilities.sort(function (a, b) {
+    mtgAbilities: mtgAbilities.sort(function (a, b) {
       if (a.nameFr < b.nameFr) {
         return -1;
       }
@@ -42,7 +42,7 @@ export default new Vuex.Store({
         return 1;
       }
       return 0;
-    }) as MagicAbility[],
+    }) as MtgAbility[],
 
     scythePlays: scythePlays as ScythePlay[],
     scytheFactions: scytheFactions as ScytheFaction[],